refactor(decompression): move the transform decompression context to its own header
